Update documentation. Add some steps to building jzmq to get around some wonkiness. Allow JAVA_HOME to be set outside of the script.

+++ build_jzmq.sh
@@ -8,7 +8,10 @@ fakeroot=jzmq
 origdir=$(pwd)
 description='JZMQ is the Java bindings for ZeroMQ'
 # Make sure JAVA_HOME is set. Uncomment if necessary
-export JAVA_HOME='/usr/lib/jvm/java-6-sun-1.6.0.31'
+if [ "${JAVA_HOME}x" == "x" ]; then
+  echo Please set JAVA_HOME before running.
+  exit -1
+fi
 
 #_ MAIN _#
 rm -rf ${name}*.deb
@@ -19,6 +22,12 @@ git clone ${url}
 cd jzmq
 ./autogen.sh
 ./configure
+
+cd src/
+touch classdist_noinst.stamp
+CLASSPATH=.:./.:$CLASSPATH javac -d . org/zeromq/ZMQ.java org/zeromq/App.java org/zeromq/ZMQForwarder.java org/zeromq/EmbeddedLibraryTools.java org/zeromq/ZMQQueue.java org/zeromq/ZMQStreamer.java org/zeromq/ZMQException.java
+cd ..
+
 make
 make install DESTDIR=${origdir}/${buildroot}
